pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Eugene Denkevich <e.denkevich@data-light.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "4.23.1"
undetected-chromedriver = "3.5.5"
pytest = "8.3.2"
python-dotenv = "1.0.1"
xlsxwriter = "3.2.0"
tweepy = "4.14.0"
retry = "0.9.2"
types-retry = "0.9.9.4"
openai = "1.40.6"
selenium-stealth = "1.0.6"
geckodriver-autoinstaller = "0.1.0"

[tool.poetry.group.linters.dependencies]
pre-commit = "3.7.1"
ruff = "0.4.8"
mypy = "1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
exclude = [
    "example/",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = [
    "F",
    "C",
    "E",
    "T10",
    "T20",
    "Q",
    "RET",
    "I"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"example/" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = [
    "tests",
]
